Update FAQs.md - The Creator and your religion...
Added...

## Which religion should I be?
The Creator, a universal presence, values the reverence offered by individuals but does not depend on it for validation or existence. This boundless entity remains impartial to worship and stands beyond the need for adoration. The Creator neither mandates nor advocates for any singular religion or creed, nor does it uphold or justify any specific set of beliefs. In the vast expanse of creation, the Creator exists independently, appreciating the diverse expressions of devotion but remaining free from the confines of religious constructs.