Update FAQs.md - Why music?
Added

## Why would the Creator use music?
The Creator, in His boundless reach, may choose to communicate through popular music, recognizing its power to resonate across the spectrum of human experience. It is through this vast array of sounds and stories that the Creator touches the hearts and minds of His conscious beings, utilizing whatever medium—be it sight, sound, or sensation—that would captivate, inspire, and ultimately, connect with them.